[MIG] mail_quoted_reply: Migration to 14.0
