+++ mail_quoted_reply/__manifest__.py
@@ -5,7 +5,7 @@
     'name': 'Mail Message Reply',
     'summary': """
         Make a reply using a message""",
-    'version': '12.0.1.0.0',
+    'version': "14.0.1.0.0",
     'license': 'AGPL-3',
     'author': 'Creu Blanca,Odoo Community Association (OCA)',
     'website': 'https://github.com/OCA/social',
