+++ mail_quoted_reply/static/src/scss/thread_mail_message_reply.scss
@@ -1,0 +1,21 @@
+.o_thread_mail_message_reply {
+  cursor: pointer;
+  color: gray('400');
+  &:not(.o-mobile) {
+        &:hover {
+            filter: brightness(0.8);
+        }
+    }
+
+    &.o-mobile {
+        filter: brightness(0.8);
+
+        &:hover {
+            filter: brightness(0.75);
+        }
+    }
+
+    &.o-message-selected {
+        color: gray('500');
+    }
+}
