+++ mail_quoted_reply/static/src/xml/mail_message_reply.xml
@@ -2,11 +2,18 @@
 <!-- Copyright 2021 Creu Blanca
      License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html). -->
 <template>
-    <t t-extend="mail.widget.Thread.Message">
-        <t t-jquery=".o_thread_message_reply" t-operation="after">
-            <i t-if="message.isLinkedToDocumentThread() and !options.displayReplyIcons and !message.isSystemNotification()"
-               class="fa fa-reply o_thread_icon o_thread_mail_message_reply"
-               t-att-data-message-id="message.getID()" title="Reply" role="img" aria-label="Reply"/>
-        </t>
+    <t t-inherit="mail.Message" t-inherit-mode="extension">
+        <xpath
+            expr="//div[hasclass('o_Message_core')]//div[hasclass('o_Message_headerCommands')]/t[1]"
+            position="inside"
+        >
+            <t>
+                <span
+                    class="fa fa-reply o_thread_icon o_thread_mail_message_reply"
+                    t-on-click="_onClickMailMessageReply"
+                    title="Reply"
+                />
+            </t>
+        </xpath>
     </t>
 </template>
